YAN-49 Настройка серверной инфраструктуры
