+++ packages/server/db.ts
@@ -1,28 +1,33 @@
-import { Client } from 'pg'
+import { Sequelize, SequelizeOptions } from 'sequelize-typescript'
 
-const { POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT }
-  = process.env
+const { 
+  POSTGRES_USER, 
+  POSTGRES_PASSWORD, 
+  POSTGRES_DB, 
+  POSTGRES_PORT, 
+  POSTGRES_HOST 
+} = process.env
 
-export const createClientAndConnect = async (): Promise<Client | null> => {
+const host = process.env.NODE_ENV === 'development' ? 'localhost' : POSTGRES_HOST
+
+const sequelizeOptions: SequelizeOptions = {
+  host: host,
+  port: Number(POSTGRES_PORT),
+  username: POSTGRES_USER,
+  password: POSTGRES_PASSWORD,
+  database: POSTGRES_DB,
+  dialect: 'postgres',
+  models: []
+}
+
+export const sequelize = new Sequelize(sequelizeOptions)
+
+export async function dbConnect() {
   try {
-    const client = new Client({
-      user: POSTGRES_USER,
-      host: 'localhost',
-      database: POSTGRES_DB,
-      password: POSTGRES_PASSWORD,
-      port: Number(POSTGRES_PORT),
-    })
-
-    await client.connect()
-
-    const res = await client.query('SELECT NOW()')
-    console.log('  ➜ 🎸 Connected to the database at:', res?.rows?.[0].now)
-    client.end()
-
-    return client
-  } catch (e) {
-    console.error(e)
+    await sequelize.authenticate()
+    await sequelize.sync()
+    console.log('👍 Соединение с БД успешно установлено')
+  } catch (error) {
+    console.error('🚨 Ошибка при подключении в БД', error)
   }
-
-  return null
 }
